增加疫情地图功能

+++ community-logging/src/main/java/com/example/csl/controller/GetEmailCodeController.java
@@ -1,17 +1,20 @@
 package com.example.csl.controller;
 
-import com.example.csl.Service.Impl.RedisServiceImpl;
-import com.example.csl.Utils.IdentifyCodeUtil;
+import com.example.csl.service.Impl.RedisServiceImpl;
+import com.example.csl.utils.IdentifyCodeUtil;
+import com.example.csl.bean.FsEmail;
 import com.example.csl.result.Result;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.mail.SimpleMailMessage;
 import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.validation.BindingResult;
 import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RestController;
 
 import javax.annotation.Resource;
+import javax.validation.Valid;
 
 /**
  * @description: 发送邮件获取验证码
@@ -31,14 +34,23 @@ public class GetEmailCodeController {
 
     //发送普通文本邮件
     @PostMapping(value = "/user/email")
-    public Result sendMail(String email) {
+    public Result sendMail(@RequestBody @Valid FsEmail fsEmail,BindingResult results) {
+        Result result = new Result();
+        if (results.hasErrors()) {
+            result.setStatus(0);
+            result.setMessage(results.getFieldError().getDefaultMessage());
+            return result;
+        }
+        String email = fsEmail.getEmail();
 
+
+        System.out.println(email);
         if (redisService.getValue(email)!=null){
             redisService.delete(email);
         }
 
         SimpleMailMessage message = new SimpleMailMessage();
-        Result result = new Result();
+
 
         String code = IdentifyCodeUtil.getRandom();
         message.setFrom("279126594@qq.com"); //发送者
@@ -53,6 +65,8 @@ public class GetEmailCodeController {
             result.setStatus(0);
             result.setMessage("Failed to get the verification code.");
         }
+
+
 
         return result;
     }
